requirements.txt
pip
Flask
requests
cassandra-driver
passlib
flask-login
